Restify the collections APIs
